Refactoring 1
Corrections in the routes, corrections in the readme and changing the type of the amount of float to integer

+++ phoenix/app/cad_products_phoenix/lib/cad_products_phoenix/management/register.ex
@@ -13,7 +13,7 @@ defmodule CadProductsPhoenix.Management.Register do
     field :name, :string
     field :description, :string
     field :price, :float
-    field :qtd, :float
+    field :qtd, :integer
   end
 
   @doc false
